backend/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
bidict==0.21.4
cachetools==5.0.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
Deprecated==1.2.13
dill==0.3.4
easydict==1.9
ffmpeg-progress-yield==0.2.0
ffmpeg-quality-metrics==2.3.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gevent==21.12.0
gevent-websocket==0.10.1
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.0
greenlet==1.1.2
grpcio==1.32.0
h11==0.13.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
itsdangerous==2.1.1
Jinja2==3.0.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.2.1
numpy==1.19.5
oauthlib==3.2.0
opencv-python==4.1.1.26
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
promise==2.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.2
pytz==2022.1
redis==4.2.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
simple-websocket==0.5.1
six==1.15.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.1
tensorflow-datasets==4.3.0
tensorflow-estimator==2.4.0
tensorflow-metadata==1.7.0
termcolor==1.1.0
tqdm==4.63.0
typing-extensions==3.7.4.3
urllib3==1.26.9
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.12.1
wsproto==1.1.0
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
